Fix icon view route collision with IronIconsAddon Moved redirect classes to dedicated package
